refactor: move fsUtils to shared utils location (#3768)
Move general-purpose filesystem utilities (fileExistsAt, readFileText)
from interpreters/ to utils/ since they are used across unrelated modules.
