+++ extensions/vscode/src/utils/fsUtils.test.ts
