+++ extensions/vscode/src/utils/fsUtils.ts
